Fix build and test for ubuntu builds
This fixes the ubuntu builds and tests an initial sanity test of 'is
eegnb installed correctly?'

Changes:
- use linux for 'runner.os' query string
- create virtual display for Pyglet dependencies in PsychoPy
- add freeglut3-dev package for PyschoPY transitive deps
- install wxPython, `pyo`, and `h5py` python packages
- add two 'sanity' test running `eegnb`
- add packages required for linux build in requirements.txt
- build docs on macOS to avoid having to install wxPython

Note:
wxPython is required for Linux bulids and takes ~35 minutes to build on
the current CI Runner (as well as locally). We can optimize in the
future by building + storing the wxPython egg.