Fix build and test for ubuntu builds
This fixes the ubuntu builds and tests an initial sanity test of 'is
eegnb installed correctly?'

Changes:
- use linux for 'runner.os' query string
- create virtual display for Pyglet dependencies in PsychoPy
- add freeglut3-dev package for PyschoPY transitive deps
- install wxPython, `pyo`, and `h5py` python packages
- add two 'sanity' test running `eegnb`

There is room for improvement, for example, the python packages need to
be added to the requirements.txt file, but it currently takes 30 minutes
to build wxPython. We need to find a way to optimize that first,
probably (by building an easy to distribute pre-built binary of the
package for this specific runner.os).