Update dependencies and fix compatibility issues for macOS Apple Silicon
This commit introduces multiple changes that ensure compatibility with macOS Apple Silicon. The 'keyboard' dependency was removed because of segmentation faults, and it was replaced with 'pynput'. The versions of 'pylsl', 'pyobjc', and 'psychopy' were specified to prevent dependency resolution loop and solve issues on Apple Silicon. The 'conda' installation step was added to the GitHub Actions workflow to facilitate installation of these dependencies. Some experiment and analytical scripts were also updated accordingly for compatibility.